requirements.txt.txt
streamlit
tensorflow
numpy
Pillow
opencv-python